requirements.txt
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
APScheduler==3.6.3
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==4.2.2
certifi==2025.8.3
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.2.1
curl_cffi==0.13.0
cycler==0.12.1
ephem==4.1.6
et-xmlfile==1.1.0
exceptiongroup==1.1.1
flatlib==0.2.3
fonttools==4.51.0
frozendict==2.4.4
google-auth==2.42.0
google-genai==1.47.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.28.1
idna==3.4
kiwisolver==1.4.5
lxml==5.2.1
matplotlib==3.8.4
mplfinance==0.12.10b0
multitasking==0.0.11
numpy==1.24.3
openpyxl==3.1.2
outcome==1.2.0
packaging==24.0
pandas==2.0.1
peewee==3.17.3
pillow==10.3.0
platformdirs==4.4.0
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.1.2
pyperclip==1.8.2
pypnf==0.0.7
PySocks==1.7.1
pyswisseph==2.8.0.post1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.1.1
python-telegram-bot==13.13
pytz==2023.3
requests==2.31.0
rsa==4.9.1
scipy==1.13.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
ta==0.11.0
tabulate==0.9.0
tenacity==9.1.2
tornado==6.4
trio==0.22.0
trio-websocket==0.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.15
webdriver-manager==4.0.2
webencodings==0.5.1
websockets==15.0.1
wsproto==1.2.0
XlsxWriter==3.1.1
yfinance==0.2.65
